Restaurama class PoC: capability of fetching all restaurants in the DB and interactive CLI enablement
