+++ .idea/compiler.xml
@@ -6,10 +6,11 @@
         <sourceOutputDir name="target/generated-sources/annotations" />
         <sourceTestOutputDir name="target/generated-test-sources/test-annotations" />
         <outputRelativeToContentRoot value="true" />
-        <module name="restaurant-service" />
-        <module name="order-service" />
-        <module name="payment-service" />
+        <module name="orders-service" />
         <module name="common" />
+        <module name="users-service" />
+        <module name="restaurants-service" />
+        <module name="payments-service" />
       </profile>
     </annotationProcessing>
   </component>
