+++ .idea/encodings.xml
@@ -3,11 +3,11 @@
   <component name="Encoding">
     <file url="file://$PROJECT_DIR$/common/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/common/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/order-service/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/order-service/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/payment-service/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/payment-service/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/restaurant-service/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/restaurant-service/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/orders-service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/orders-service/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/payments-service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/payments-service/src/main/resources" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/restaurants-service/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/restaurants-service/src/main/resources" charset="UTF-8" />
   </component>
 </project>
