+++ .idea/sqldialects.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/mysql-init-seed-data.sql" dialect="MySQL" />
+    <file url="file://$PROJECT_DIR$/db-init-resources/mysql/mysql-init-seed-data.sql" dialect="MySQL" />
     <file url="PROJECT" dialect="MySQL" />
   </component>
 </project>
