+++ pom.xml
@@ -13,9 +13,10 @@
     <url>https://github.com/AliAS-86/restaurama-food-app</url>
     <modules>
         <module>common</module>
-        <module>restaurant-service</module>
-        <module>order-service</module>
-        <module>payment-service</module>
+        <module>restaurants-service</module>
+        <module>orders-service</module>
+        <module>users-service</module>
+        <module>payments-service</module>
 
     </modules>
     <!-- Maven Compiler Plugin - updating the source and target to use Java 17 -->
