+++ restaurants-service/pom.xml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <!-- Reference to the parent POM -->
+    <parent>
+        <groupId>com.restaurama</groupId>
+        <artifactId>restaurama-food-app</artifactId>
+        <version>1.0.0-SNAPSHOT</version>
+    </parent>
+
+    <!-- Service artifact -->
+    <artifactId>restaurants-service</artifactId>
+    <!-- Adding module specific dependencies -->
+    <dependencies>
+        <dependency>
+            <groupId>com.restaurama</groupId>
+            <artifactId>common</artifactId>
+            <version>1.0.0-SNAPSHOT</version>
+        </dependency>
+    </dependencies>
+
+</project>
